backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = """
This service fetches data from an Analytics Service like Google Analytics,
cleans it and transforms it to a format for visualisation.
Also provides REST endpoints for the frontend.
"""
authors = ["Roopak Nijhara <roopak@parallelhq.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.1"
google-api-python-client = "^2.39.0"
oauth2client = "^4.1.3"
python-dotenv = "^0.19.2"
SQLAlchemy = "^1.4.32"
psycopg2 = "^2.9.3"
requests = "^2.27.1"
google-auth = "^2.6.2"
python-dateutil = "^2.8.2"
google-analytics-data = "^0.11.2"
google-auth-oauthlib = "^0.5.1"
boto3 = "^1.23.8"
fastapi = {extras = ["all"], version = "^0.78.0"}
uvicorn = "^0.17.6"
beanie = "^1.11.8"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
pytest-mock = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
